==> scripts/ppp_Jenny/scripts/06_fiber_alpha_abstract.Rmd <==
---
title: "human fiber and alpha abstract"
author: "Angel"
date: "2022-11-16"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse) 
library(ggpubr)         
```

```{r}
all <- read_csv('~/pipeline/scripts/food_tree/data/090_all_samples_meta_p2d_fg9_dietall_genera34.csv') %>% 
  mutate(ave_fiber = ave_fiber*100) 
meta <- read_csv('~/pipeline/scripts/food_tree/data/090_all_samples_meta_p2d_fg9_dietall_genera34.csv') %>% 
  select(mrn, sdrt, simpson_reciprocal, ave_fiber) %>% 
  filter(sdrt %in% 7:21)

meta %>% distinct(mrn)
```

Maybe look into GVHD yes/no (on x axis) and averaged fiber intake (on the y axis- this has to be one data point per patient)?
Another idea is to have a plot with the x-axis being time from transplant and the y axis fiber intake and color points by GVHD status (yes vs no) and add a line (geom_smooth) for GVHD yes vs no ?
If that doesn’t separate well you can also look only into the lower GI GVHD vs no GVHD (remove skin, liver and upper GI GVHD).
The other idea is to look into gvhd-related mortality: divide patients in high vs low fiber (i would use data from peri-engraftment only) and the GVHD related mortality as a landmark analysis - starting at day 21 (end of peri-engraftment period) and up to 2 years. 
Just some ideas.. and hopefully that will give you some interesting results!!

```{r}
# find the gvhd data of the patients
library(vdbR)
connect_database()
get_table_from_database('patient_allo_ks_20221104')
ptb <- read_csv('~/pipeline/scripts/food_tree/data/cleaned_patients/diet_patients_97.csv') %>% 
  filter(mrn %in%  meta$mrn)
gvhd <- patient_allo_ks_20221104 %>% 
  filter(mrn %in% meta$mrn) %>% 
  distinct(mrn, gvhd_old, hct, .keep_all = T ) %>% 
  inner_join(ptb, by = c("mrn", "hct"))

combined <- meta %>% 
  left_join(gvhd, by = "mrn")

combined %>% write_csv('../data/06_for_abstract_peri.csv')

combined %>% 
  group_by(mrn,gvhd_old ) %>% 
  summarise(mean_fiber = mean(ave_fiber)) %>% 
  ggboxplot(x = 'gvhd_old', y = 'mean_fiber', add = 'jitter') +
  stat_compare_means(comparisons= list(c('N', 'Y')),
            label= "p.format",
            method= 'wilcox.test',
						correct=FALSE)

ggsave('../data/06_gvhd_mean_fiber.pdf', height = 5)
```


```{r}
meta %>% 
  group_by(mrn) %>% 
  summarise(ave_div = mean(simpson_reciprocal),
            mean_fiber = mean(ave_fiber)) %>% 
  ggscatter(x = 'mean_fiber', y = 'ave_div',
            xlab = 'Prior 2 day fiber intake averaged by patient', ylab = 'microbiome alpha diversity averaged by patient',
            add = "reg.line",  # Add regressin line
           add.params = list(color = "blue", fill = "lightgray"), # Customize line
           conf.int = TRUE, # Add confidence interval
           cor.coef = TRUE, # Add correlation coefficient.
           cor.coeff.args = list(method = "pearson",  label.sep = "\n")) +
  theme(aspect.ratio=1)

ggsave('../data/06_diversity_and_fiber.pdf', height = 5)
```

```{r}
# a box plot
mean_fiber_intake <- mean(meta$ave_fiber)
median(meta$ave_fiber)
grps <- meta %>% 
  mutate(grp = if_else(ave_fiber > mean_fiber_intake, 'high', 'low'))

grps %>% 
  ggboxplot(x = 'grp', y = 'simpson_reciprocal', xlab = 'Compared to average prior 2 day fiber intake',
            ylab = 'Microbiome alpha diversity') +
  stat_compare_means(comparisons= list(c('high', 'low')),
            label= "p.format",
            method= 'wilcox.test',
						correct=FALSE)

ggsave('../data/06_diversity_and_fiber_binary.pdf', height = 5)
```


For patients: could you do fiber over time? y-axis: fibers; x-axis time as a continuous (don't break it as pre-transplant, perengraftment etc) and color by GVHD vs controls?

```{r}
# find all of th gvhd or not for the patients
full <- read_csv('../../food_tree/data/090_all_samples_meta_p2d_fg9_dietall_genera34.csv')
pts <- full %>% 
  distinct(mrn)

gvhd <- patient_allo_ks_20221104 %>% 
  filter(mrn %in% pts$mrn) %>% 
  distinct(mrn, gvhd_old, hct, .keep_all = T ) %>% 
  inner_join(ptb, by = c("mrn", "hct")) %>% 
  select(mrn,  gvhd_old) %>% 
  filter(gvhd_old %in% c('N','Y'))
gvhd %>% distinct(mrn)
gvhd %>% count(gvhd_old)

df <- full %>% 
  select(mrn, sampleid, sdrt, ave_fiber) %>% 
  mutate(ave_fiber = ave_fiber *100) %>% 
  inner_join(gvhd )

df %>% 
  ggscatter(x  = 'sdrt', y = 'ave_fiber', color = 'gvhd_old', palette = 'nejm', add = "loess", conf.int = T,
            alpha = 0.7, 
            ylab = 'average fiber intake in previous two days',
            xlab = 'day relative to transplant')

ggsave('../data/06_fiber_versus_time.pdf', width = 5, height = 4)
```
