==> scripts/ppp_Jenny/scripts/08_fiber_rich_foods.Rmd <==
---
title: "fiber rich foods"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

when patients are eating higher fiber what are they eating

day 0-21

```{r}
library(tidyverse)
library(ggpubr)
key <- read_csv('../../food_tree/data/cleaned_diet_data/food_group_color_key_final.csv', col_types = 'ccccc')
```

```{r}
dtb1 <- read_csv('../../food_tree/data/149_finalized_dtb1.csv')
dtb2 <- read_csv('../../food_tree/data/149_finalized_dtb2.csv')
unit <- read_csv('../../food_tree/data/148_both_batches_UNIT_table_EN_exclu.csv')
DTB <- bind_rows(dtb1, dtb2)
colnames(unit)
```

```{r}
# distribution of fiber psu
summary(unit$Fiber_psu)

# the percentage of the fiber in dry weight
perc <- unit %>% 
  mutate(fiber_dry_perc = round(Fiber_psu/dehydrated_psu*100, 2))

# the distribution of the above
summary(perc$fiber_dry_perc)

# the foods that is above the median perc
above2.38 <- perc %>% 
  filter(fiber_dry_perc >= median(perc$fiber_dry_perc)) %>% 
  select(food_all, Food_code, description, Fiber_psu, fiber_dry_perc) %>% 
  distinct(Food_code, description, fiber_dry_perc) %>% 
  arrange(desc(fiber_dry_perc)) %>% 
  mutate(fgrp1 = str_sub(as.character(Food_code), 1, 1)) %>% 
  left_join(key %>% select(fgrp1, fdesc)) %>% 
  mutate(fdesc = fct_reorder(fdesc, fiber_dry_perc, .fun=median, .desc = T))

# make box plot to see the groups and the fiber perc distribution
above2.38 %>% 
  ggboxplot(x = 'fdesc', y = 'fiber_dry_perc', orientation = 'horizontal',
            add = 'jitter', 
            ylab = 'fiber_g/dry_weight_g %')

above2.38 %>% distinct(fgrp1)
# Fats, Oils & Salad Dressings does not have fiber 

# the things in the group 9 that is high in fiber
above2.38 %>% 
  filter(fgrp1 == '9') %>% 
  arrange(desc(fiber_dry_perc))
```
```{r}
# the patients intake of these foods 
# the patients fiber intake during the days 0-21
fiber_per_day <- DTB %>% 
  group_by(mrn, fdrt) %>% 
  summarise(fiber_pd = sum(Fibers_g))

fiber_per_day %>% 
  ggscatter(x = 'fdrt', y = 'fiber_pd', size = 0.5) +
  geom_line(aes(group = mrn), color = "grey", alpha = 0.5) +
  facet_wrap(~ mrn ) 
ggsave('../data/08_pt_fiber.pdf', width = 15, height = 15)

```

```{r}
# Can you calculate the fiber intake in % per patient and per day, as well as the absolute grams? This way we may be able to inform the mice experiments
# calculate the dry weight of food intake per pt per day
dry_perday <- DTB %>% 
  group_by(mrn, fdrt) %>% 
  summarise(dry_pd = sum(dehydrated_weight))

fiber_perc <- fiber_per_day %>% 
  full_join(dry_perday, by = c("mrn", "fdrt")) %>% 
  mutate(fiber_percent = round(fiber_pd/dry_pd*100, 2))

fiber_perc %>% write_csv('../data/patients_16s_data/08_patients_fiber_absweight_and_percent.csv')
```


```{r}
# day 0-21
fiber_per_day %>% 
  filter(fdrt %in% 0:21) %>% 
  ggscatter(x = 'fdrt', y = 'fiber_pd', size = 0.5) +
  geom_line(aes(group = mrn), color = "grey", alpha = 0.5) +
  facet_wrap(~ mrn ) 
ggsave('../data/08_pt_fiber_day-21.pdf', width = 15, height = 15)
```

```{r}
# see the distribution of the per day fiber eating
summary(fiber_per_day$fiber_pd)

# median for the day 0-21
fiber_per_day_day21 <- fiber_per_day %>% 
  filter(fdrt %in% 0:21)

summary(fiber_per_day_day21$fiber_pd)

# see what the foods are for the patients that had above 75% fiber 

above0.75fiber <- fiber_per_day_day21 %>% 
  filter(fiber_pd >= quantile(fiber_per_day_day21$fiber_pd, 0.75))

# how many patients 
above0.75fiber %>%  distinct(mrn)

# order by the number of days that are high fiber days
above0.75fiber_days <- above0.75fiber %>% 
  count(mrn, sort = T) 

summary(above0.75fiber_days$n)

# at least 4 days of high fiber days patients 
highfiber_4day <- above0.75fiber_days %>% 
  filter(n >= 4)

pt_order <- highfiber_4day %>% 
  pull(mrn)
```

```{r}
# - can you classify patients based on how many days they had high fiber diet (between days 0 and 21) in 3 groups:
#     - High fiber: at least 4 days of high fiber diet
#     - Mod fiber: 1-3 days of high fiber diet
#     - Low fiber: no high fiber diet
all_patients_groups <- above0.75fiber_days %>% 
  full_join(DTB %>% distinct(mrn) )


all_patients_groups <- bind_rows(
  all_patients_groups %>% filter(!is.na(n)),
  all_patients_groups %>% filter(is.na(n)) %>% mutate(n = 0)
) 

all_patients_groups_fiber <- all_patients_groups %>% 
  mutate(fiber_grp = case_when(
    n >= 4 ~ 'High',
    n %in% 1:3 ~ 'Mod',
    n == 0 ~ 'Low'
  ))
```


```{r}
# find the gvhd of the patients 
library(vdbR)
connect_database()
get_table_from_database('patient_allo_ks_20221104')


pt_clinincal <- patient_allo_ks_20221104 %>% 
  filter(mrn %in% DTB$mrn) %>% 
  select(mrn, gvhd_old, d100_a_gvhd, hct)


```


```{r}
# what food groups are the high fiber coming from for those patients
highfiber_pt <- DTB %>% 
  filter(mrn %in% highfiber_4day$mrn & fdrt %in% 0:21) %>% 
  mutate(fgrp1 = str_sub(as.character(Food_code), 1, 1)) %>% 
  left_join(key %>% select(fgrp1, fdesc))  %>% 
  mutate(mrn = factor(mrn, levels =pt_order )) %>% 
  mutate(fdesc = str_sub(fdesc , start = 1,end = 6))

highfiber_pt %>% 
  ggboxplot(x = 'fdesc', y = 'Fibers_g', orientation = 'horizontal',color = 'fdesc', 
            add = 'jitter', 
            ylab = 'Fibers_g') +
  facet_wrap(~ mrn , ncol = 6 , scales = 'free' ) 
ggsave('../data/08_highfiber_pt_fg.pdf', width = 15, height = 25)
```

```{r}
# the top 5 patients
top5 <- all_patients_groups_fiber %>% 
  ungroup() %>% 
  slice_max(order_by = n, n = 5)

DTB %>% 
  filter(mrn %in% top5$mrn & fdrt %in% 0:30) %>% 
  mutate(fgrp1 = str_sub(as.character(Food_code), 1, 1)) %>% 
  left_join(key %>% select(fgrp1, shortname))  %>% 
  group_by(mrn, fdrt, shortname) %>% 
  summarise(daily_fiber = sum(Fibers_g)) %>% 
  mutate(shortname = fct_reorder(shortname, daily_fiber, .fun=median, .desc = F)) %>% 
  #mutate(mrn = factor(mrn, levels =pt_order )) %>% 
  ggboxplot(x = 'shortname', y = 'daily_fiber', orientation = 'horizontal',color = 'shortname', 
            add = 'jitter', xlab = '',
            ylab = 'Daily fiber intake from each food groups')
ggsave('../data/08_Daily fiber intake from each food groups_top5.pdf')
ggsave('../data/patients_16s_data/08_Daily fiber intake from each food groups_top5.pdf')
```

```{r}
# to get the patient info of 
# -Median age
# -Male/Female distribution
# -Race and Ethnicity
# -Hematological diagnosis
# get the race from the original table
ptb  <-  read_csv('../data/11_ptb_characteristics.csv')
library(vdbR)
connect_database()
get_table_from_database('patient_allo_ks_20221104')
race <- patient_allo_ks_20221104 %>% 
  filter(mrn %in% pts$mrn) 

ptbr <- ptb %>% 
  left_join(race %>% select(mrn, race_dscrp,ethnicity ))
  
ptbr %>% write_csv('../data/11_ptb_characteristics.csv')

```
