==> scripts/ppp_Jenny/scripts/09_fiber_types.Rmd <==
---
title: "Fiber types in foods"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(ggpubr)
```

```{r}
types <- readxl::read_excel('../data/148_both_batches_UNIT_table_EN_exclu_UPDATEDJP (1).xlsm', sheet = 3) %>% 
  mutate(fraction = Percentage/100) %>% 
  mutate(Food = str_to_lower(Food),
         Food = str_replace(Food, 's$',''),
         Food = str_replace(Food, '_$',''))

DTB <- read_csv('../../food_tree/data/152_combined_DTB.csv')

DTB %>% 
  distinct(mrn, fdrt)
```

looking at cellulose over time in patients with GVHD vs no GVHD (geom_point and geom_smooth, colored by GVHD day -30 to +30)
similarly looking at soluble fibers over time in patients with GVHD vs no GVHD (geom_point and geom_smooth, colored by GVHD day -30 to +30)
similarly looking at insoluble fibers over time in patients with GVHD vs no GVHD (geom_point and geom_smooth, colored by GVHD day -30 to +30)

```{r}
# find how many food_nsc have each of the "Food"
unique_food <- types %>% 
  distinct(Food) %>% pull(Food)

unique_foodnsc <- DTB %>% 
  mutate(Food_NSC = str_to_lower(Food_NSC)) %>% 
  distinct(Food_NSC, Food_code, description) 

found <- unique_food %>% 
  set_names(unique_food) %>% 
  map(function(food_str){
    unique_foodnsc %>% 
      filter(str_detect(Food_NSC, food_str))
  })

# the foods that have no matches
notfound <- found %>% 
  keep( ~ nrow(.) == 0)

# names of the not found food
names(DTB)
```

```{r}
# to find the unique description/food_code for the fiber subtype annotation
toannotate <- DTB %>% 
  distinct(Food_code, description, Fibers_g) %>% 
  group_by(Food_code, description) %>% 
  arrange(desc(Fibers_g), .by_group = T) %>% 
  slice(1)

anyfiber <- toannotate %>% 
  filter(Fibers_g > 0)

# unit table
unit <- read_csv('../../food_tree/data/148_both_batches_UNIT_table_EN_exclu.csv')

unit %>% distinct(Food_code)

DTB %>% filter(Meal == 'Enteral nutrition') %>% distinct(Food_code)

setdiff(toannotate$Food_code, unit$Food_code)

# find the things that have the 

units <- unit %>% 
  distinct(Food_code, description, Fiber_psu, dehydrated_psu, .keep_all = T) %>% 
  group_by(Food_code, description) %>% 
  arrange(desc(Fiber_psu), .by_group = T) %>% 
  slice(1) %>% 
  mutate(fiber_perc = round(Fiber_psu/dehydrated_psu*100, 2)) %>% 
  filter(fiber_perc > 0) %>% 
  select(Food_NSC:description, Fiber_psu, fiber_perc) %>% 
  arrange(Food_code)

# add the 5 EN ones
en_unit <- DTB %>% filter(Meal == 'Enteral nutrition') %>% 
  mutate(description = if_else(Food_NSC == 'Suplena 1.8', 'Enteral formula, suplena', description)) %>% 
  distinct(Food_NSC,Unit, Food_code, description, .keep_all = T) %>% 
  mutate(Fiber_psu = Fibers_g/Por_eaten,
         dehydrated_psu = dehydrated_weight/Por_eaten) %>% 
  mutate(fiber_perc = round(Fiber_psu/dehydrated_psu*100, 2)) %>% 
  select(Food_NSC, Unit, Food_code,description, Fiber_psu, fiber_perc) %>% 
  arrange(Food_NSC) %>% 
  filter(fiber_perc > 0)

all_units <- bind_rows(en_unit, units) %>% arrange(Food_code)

all_units %>% write_csv('../data/09_unit_serving_food_fiber.csv')
```


```{r}
# ok Jenny sent a table, found the missing ones
jenny <- readxl::read_excel('../data/148_both_batches_UNIT_table_EN_Fiber_Types_JP.xlsm', sheet = 3)

# find the missing ones
length(intersect(jenny$Food_code, all_units$Food_code))

setdiff( all_units$Food_code, jenny$Food_code)

need <- all_units %>% 
  filter(! Food_code %in% jenny$Food_code) %>% 
  arrange(Food_code)

need %>% write_csv('../data/09_more_need_info.csv')
```
