==> scripts/ppp_Jenny/scripts/11_analysis_patients_fiber_type.Rmd <==
---
title: "Analysis of the fiber types of the patients"
author: "Anqi Dai"
date: "`r Sys.Date()`"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse) 
library(ggpubr)
library(vdbR)
connect_database()
list_table_from_database('fmt')
get_table_from_database('fmt_arm_ag')
```

-Calculate the population median of fiber intake (days -10 to +30), from all the patients. Including zeros. 
-Based on the population median, calculate the high and low fiber intake distribution (all patients combined).
-Calculate the high and low fiber distribution in GVHD vs. No-GVHD patients.
-Calculate the absolute fiber intake in grams for each fiber type (days -10 to +30), allocating values of zero for the remaining fiber types, in GVHD and No-GVHD patients.  

```{r}
type <- read_csv('../data/10_patients_foods_fiber_type_all.csv')
DTB <- read_csv('../../food_tree/data/152_combined_DTB.csv')
# join with gvhd info of the patients
allpt <- read_csv('../../food_tree/data/156_transplant_data.csv') 
allpt %>% count(source)

allpt_gvhd <- allpt %>% 
  mutate(d100_a_gvhd = if_else(d100_a_gvhd == 'NE', 'N/E', d100_a_gvhd),
         d100_a_gvhd = if_else(d100_a_gvhd == 'N', 'No', d100_a_gvhd),
         d100_a_gvhd = if_else(d100_a_gvhd == 'Y', 'Yes', d100_a_gvhd)) %>% 
  filter(d100_a_gvhd %in% c('Yes','No')) 

tcd_pt <- allpt %>% 
  filter(source == 'TCD') %>% 
  pull(mrn)

meta <- read_csv('../../food_tree/data/153_combined_META.csv')
 
allpt_gvhd_all <- allpt_gvhd %>% 
  filter(source != 'TCD') %>% 
  mutate(grp = case_when(
    d100_a_gvhd == 'No' ~ "no_gvhd",
  d100_a_gvhd == 'Yes' & d100_upper_gi_stage == 0 & d100_lower_gi_stage == 0 ~ "non_gi",
  d100_a_gvhd == 'Yes' & d100_upper_gi_stage != 0 & d100_lower_gi_stage == 0  ~ "upper_gi",
  d100_a_gvhd == 'Yes' &  d100_lower_gi_stage != 0  ~ "LGI",
  .default = NA
  )) %>% 
  select(mrn, source, grp, hct) %>% 
  inner_join(meta %>% select(mrn, sampleid, sdrt)) %>% 
  left_join(fmt_arm_ag %>% select(mrn, fmt_procedure_date)) %>% 
  mutate(hct = ymd(hct),
         fmtday = as.numeric(fmt_procedure_date - hct)) %>% 
  mutate(grp = if_else(grp %in% c('non_gi','upper_gi'), 'Non_LGI', grp)) %>% 
  # since there are no samples that are after FMT we can ignore those columns
  select(-hct, -fmtday, -fmt_procedure_date)

#allpt_gvhd_all %>% write_csv('../data/11_samples_from_groups.csv')
# the below are counting stool samples
allpt_gvhd_all %>% count(grp)
```


```{r}
# calculate daily intake of each food code 
# remove the TCD patients here!!!
dailycode <- DTB %>%   
  group_by(mrn, fdrt, Food_code) %>% 
  summarise(dcodefiber = sum(Fibers_g)) %>% 
  janitor::clean_names() %>% 
  #filter(dcodefiber > 0) %>% 
  left_join(type, by = "food_code") %>% 
  ungroup() %>% 
  relocate(description, .after = 'soluble') %>% 
  inner_join(allpt_gvhd_all %>% distinct(mrn, source , grp))

dailycode %>% write_csv('../data/11_dailycode.csv')
```

```{r}
axis_text_size <- 10
axis_title_size <- 10
stip_txt_size <- 8
alpha_val <- 0.05
point_size <- 0.2
scatter_col <- 'gray14'
pick_point_size <- 1
pick_line_size <- 0.5
example_pt_line_size <- 0.3
diet_line_color <- '#E41A1C'
stool_line_color <- 'blue'
```

```{r}
dailycode %>% count(source) 

fiber_dtb <-DTB %>%   
  inner_join(allpt_gvhd_all %>% distinct(mrn, source , grp)) %>% 
  filter(fdrt %in% -10:30) 
fiber_dtb %>% distinct(mrn) %>% nrow
fiber_dtb %>% 
  distinct(mrn, Meal, fdrt) %>% nrow
fiber_dtb %>% nrow
```


```{r}
# summarize the food entries 
diet_hist <- fiber_dtb %>% 
  distinct(mrn, Meal, fdrt) %>% 
  gghistogram(x = 'fdrt', xlab = 'Transplant day', ylab = 'Meals recorded', color = diet_line_color, fill = diet_line_color, alpha = 1, bins = 60)+
    scale_x_continuous( breaks = seq(-10, 30, 10)) + 
  theme_classic(base_size = 11) +
  theme(aspect.ratio=1,
         axis.title=element_text(size=axis_title_size),
        #axis.text.x = element_blank(),
        axis.title.x = element_blank())
diet_hist
ggsave('../data/11_diet_hist.pdf', width =3, height = 3)
```
```{r}
fiber_meta <- meta %>% 
  inner_join(allpt_gvhd_all %>% distinct(mrn, source , grp)) %>% 
  filter(sdrt %in% -10:30)
fiber_meta %>% count(grp)
fiber_meta %>% distinct(mrn, grp) %>%    count(grp)
```

```{r}
stool_hist <- meta %>% 
  inner_join(allpt_gvhd_all %>% distinct(mrn, source , grp)) %>% 
  filter(sdrt %in% -10:30) %>% 
  gghistogram(x = 'sdrt', xlab = 'Transplant day', ylab = 'Stool samples', color = stool_line_color, fill = stool_line_color, alpha = 1, bins = 60) +
  scale_x_continuous( breaks = seq(-10, 30, 10)) + 

  theme_classic(base_size = 11) +
  theme(aspect.ratio=1,   axis.title=element_text(size=axis_title_size))
stool_hist  

ggsave('../data/11_stool_hist.pdf', width =3, height = 3)

meta %>% 
  inner_join(allpt_gvhd_all %>% distinct(mrn, source , grp)) %>% 
  filter(sdrt %in% -10:30) %>% nrow
```

```{r}
# the dropping trend
dailycode %>% 
  filter(fdrt %in% -10:30) %>% 
  group_by(mrn, fdrt, source) %>% 
  summarise(dayfiber = sum(dcodefiber)) %>% 
  ungroup() %>% 
  #inner_join(allpt_gvhd_all %>% select(mrn, grp)) %>% 
  ggscatter(x = 'fdrt', y = 'dayfiber', add = "loess", conf.int = TRUE, alpha = 0.05, shape = 16,
            add.params = list(color = 'forestgreen'),
            xlab = 'Transplant day', ylab = 'Daily fiber intake' ) +
  scale_y_sqrt() +theme(aspect.ratio = 1)
ggsave('../data/11_daily_all_fiber.pdf', width = 3, height = 3)
dailycode %>% distinct(grp)
```

```{r}
# Same analysis as above, but only including No-GVHD and LGI patients joined into one group)
dailycode %>% 
  filter(fdrt %in% -10:30 & grp != 'Non_LGI') %>% 
  group_by(mrn, fdrt, source) %>% 
  summarise(dayfiber = sum(dcodefiber)) %>% 
  ungroup() %>% 
  #inner_join(allpt_gvhd_all %>% select(mrn, grp)) %>% 
  ggscatter(x = 'fdrt', y = 'dayfiber', add = "loess", conf.int = TRUE, alpha = 0.05, shape = 16,
            add.params = list(color = 'forestgreen'),
            xlab = 'Transplant day', ylab = 'Daily fiber intake' ) +
  scale_y_sqrt() +theme(aspect.ratio = 1)
ggsave('../data/11_daily_all_fiber_2.pdf', width = 3, height = 3)

# number of patients
dailycode %>% 
  filter(fdrt %in% -10:30 & grp != 'Non_LGI') %>% distinct(mrn) %>% nrow
```


```{r}
# the above but with the gvhd info
cols <- c("LGI" = "red","Non_LGI" = "brown4","no_gvhd" = "darkgreen")

dailycode %>% 
  filter(fdrt %in% -10:30) %>% 
  group_by(mrn, fdrt, source, grp) %>% 
  summarise(dayfiber = sum(dcodefiber)) %>% 
  ungroup() %>% 
  ggscatter(x = 'fdrt', y = 'dayfiber', add = "loess", conf.int = TRUE, alpha = 0.1, shape = 16,
            add.params = list(color = 'grp'),
            xlab = 'Transplant day', ylab = 'Daily fiber intake' ) +
  scale_colour_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_fill_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_y_sqrt() +theme(aspect.ratio = 1)
ggsave('../data/11_daily_all_fiber_grp3.pdf', width = 4, height = 4)
```

```{r}
# the above but with the gvhd info only two groups
dailycode %>% 
  filter(fdrt %in% -10:30 & grp != 'Non_LGI') %>% 
  group_by(mrn, fdrt, source, grp) %>% 
  summarise(dayfiber = sum(dcodefiber)) %>% 
  ungroup() %>% 
  ggscatter(x = 'fdrt', y = 'dayfiber', add = "loess", conf.int = TRUE, alpha = 0.1, shape = 16,
            add.params = list(color = 'grp'),
            xlab = 'Transplant day', ylab = 'Daily fiber intake' ) +
   scale_colour_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_fill_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_y_sqrt() +theme(aspect.ratio = 1)
ggsave('../data/11_daily_all_fiber_grp2.pdf', width = 4, height = 4)
```

```{r}
# the type of the different fibers
# remove the tcd patients here!!!!!
all_types <- dailycode %>% 
  group_by(mrn, fdrt, fiber_type) %>% 
  summarise(gram = sum(dcodefiber)) %>% 
  ungroup() %>% 
  filter(!is.na(fiber_type)) %>% 
  spread('fiber_type','gram', fill = 0) %>% 
  gather('fiber_type','gram', names(.)[3]:names(.)[ncol(.)]) %>% 
  inner_join(allpt_gvhd_all %>% distinct(mrn, source, grp))

all_types %>% write_csv('../data/11_all_types.csv')
```

```{r}
all_types %>% 
  filter(fdrt %in% -10:30 ) %>% 
  ggscatter(x = 'fdrt', y = 'gram', alpha = 0.05, shape = 16,xlab = 'transplant day',
             add = "loess", conf.int = TRUE, 
            add.params = list(color = 'forestgreen', size = 1)) +
  facet_grid(~ fiber_type) +
  scale_y_sqrt()

ggsave('../data/11_fiber_all_type_daily_intake.pdf', width = 10, height = 3)
```


```{r}
# Same analysis as above, but only including the No GVHD and LGI joined into one group)
all_types %>% 
  filter(fdrt %in% -10:30 & grp != 'Non_LGI') %>% 
  ggscatter(x = 'fdrt', y = 'gram', alpha = 0.05, shape = 16,xlab = 'transplant day',
             add = "loess", conf.int = TRUE, 
            add.params = list(color = 'forestgreen', size = 1)) +
  facet_grid(~ fiber_type) +
  scale_y_sqrt()

ggsave('../data/11_fiber_all_type_daily_intake_2.pdf', width = 10, height = 3)
```


```{r}
all_types %>% 
  filter(fdrt %in% -10:30 ) %>% 
  ggscatter(x = 'fdrt', y = 'gram', alpha = 0.05, shape = 16,xlab = 'transplant day',
             add = "loess", conf.int = TRUE, 
            add.params = list(color = 'grp', size = 1)) +
  facet_grid(~ fiber_type) +
    scale_colour_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_fill_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_y_sqrt()

ggsave('../data/11_fiber_all_type_daily_intake_grp3.pdf', width = 10, height = 3)
```

```{r}
# with only the no gvhd and LGI
all_types %>% 
  filter(fdrt %in% -10:30 & grp != 'Non_LGI') %>% 
  ggscatter(x = 'fdrt', y = 'gram', alpha = 0.05, shape = 16,xlab = 'transplant day',
             add = "loess", conf.int = TRUE, 
            add.params = list(color = 'grp', size = 1)) +
      scale_colour_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_fill_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  facet_grid(~ fiber_type) +
  scale_y_sqrt()

ggsave('../data/11_fiber_all_type_daily_intake_grp2.pdf', width = 10, height = 3)
```

```{r onsetdays}
# calculat the onset day for the gvhd patients
onsetdays <-  allpt %>% distinct(mrn, hct, d100_a_gvhd_onset) %>% 
  filter(!is.na(d100_a_gvhd_onset)) %>% 
  mutate(onsetday = as.numeric(d100_a_gvhd_onset - hct))

set.seed(1)
# randomly draw one onset day from the above with replacement
pseudodays <- allpt %>% 
  distinct(mrn, hct, d100_a_gvhd_onset) %>% 
  filter(is.na(d100_a_gvhd_onset)) %>% 
  mutate(onsetday = slice_sample(onsetdays %>% select(onsetday),n = nrow(.), replace = T ) %>% pull(onsetday)) %>% 
  select(mrn, hct, onsetday)

allonset <- bind_rows(pseudodays, onsetdays %>% select(-d100_a_gvhd_onset))
```


```{r}
# the median and mean of the onset days
mean(onsetdays$onsetday)
median(onsetdays$onsetday)
# the pseudo days for the onset 
medianonset <- ceiling(median(onsetdays$onsetday))
```


```{r}
# all types but day relative to onset
all_types_o <- all_types %>% 
  inner_join(allpt %>% distinct(mrn, hct, d100_a_gvhd_onset)) %>% 
  mutate(hct = ymd(hct),
         f_date = hct + fdrt, 
         d100_a_gvhd_onset = ymd(d100_a_gvhd_onset)) %>% 
  inner_join(allonset) %>% 
  mutate(hct = ymd(hct),
         d100_a_gvhd_onset = hct + onsetday,
         d100_a_gvhd_onset = ymd(d100_a_gvhd_onset),
          fdrO = f_date - d100_a_gvhd_onset,
        fdrO = as.numeric(fdrO))

all_types_o %>% write_csv('../data/11_all_types_onset.csv')
```


```{r}
# Any chance you also annotate what the 
#median onset day would be for each cohort?
#Non-LGI GVHD
#LGI GVHD  
grp_onset <- all_types_o %>% 
  distinct(grp, hct, d100_a_gvhd_onset) %>% 
  filter(grp != 'no_gvhd') %>% 
  mutate(onsetday = as.numeric(d100_a_gvhd_onset - hct)) %>% 
  group_by(grp) %>% 
  summarise(medianonset = median(onsetday)) %>% 
  mutate(medianonset = ceiling(medianonset))
```

```{r}
# Same analysis as previous slide, only including all non-TCD patients joined into one group – relative to onset)

all_types_o %>% 
  filter(fdrt %in% -10:30 ) %>% 
  ggscatter(x = 'fdrO', y = 'gram', alpha = 0.05, shape = 16,xlab = 'day relative to onset',
             add = "loess", conf.int = TRUE, 
            add.params = list(color = 'forestgreen', size = 1)) +
  facet_grid(~ fiber_type) +
  scale_y_sqrt()

ggsave('../data/11_fiber_all_type_daily_intake_onset.pdf', width = 10, height = 3)
```

```{r}
# Same analysis as previous slide, but only including the No GVHD and LGI joined into one group – relative to onset)

all_types_o %>% 
  filter(fdrt %in% -10:30  & grp != 'Non_LGI') %>% 
  ggscatter(x = 'fdrO', y = 'gram', alpha = 0.05, shape = 16,xlab = 'day relative to onset',
             add = "loess", conf.int = TRUE, 
            add.params = list(color = 'forestgreen', size = 1)) +
  facet_grid(~ fiber_type) +
  scale_y_sqrt()

ggsave('../data/11_fiber_all_type_daily_intake_onset_2.pdf', width = 10, height = 3)
```

```{r}
all_types_o %>% 
  filter(fdrt %in% -10:30  ) %>% 
    mutate(gvhd = if_else(grp == 'no_gvhd', 'No', 'Yes'),
         gvhd = factor(gvhd, levels = c('Yes','No'))) %>% 
  ggscatter(x = 'fdrO', y = 'gram', alpha = 0.05, shape = 16,xlab = 'day relative to onset',
             add = "loess", conf.int = TRUE, 
            add.params = list(color = 'gvhd', size = 1)) +
  facet_grid(~ fiber_type) +
  scale_y_sqrt()

ggsave('../data/11_fiber_all_type_daily_intake_onset_gvhd.pdf', width = 10, height = 3)
```
```{r}
all_types_o %>% 
  filter(fdrt %in% -10:30  ) %>% 
  ggscatter(x = 'fdrO', y = 'gram', alpha = 0.05, shape = 16,xlab = 'day relative to onset',
             add = "loess", conf.int = TRUE, 
            add.params = list(color = 'grp', size = 1)) +
  facet_grid(~ fiber_type) +
        scale_colour_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_fill_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_y_sqrt()

ggsave('../data/11_fiber_all_type_daily_intake_onset_grp3.pdf', width = 10, height = 3)
```
```{r}
all_types_o %>% 
  filter(fdrt %in% -10:30 & grp != 'Non_LGI' ) %>% 
  ggscatter(x = 'fdrO', y = 'gram', alpha = 0.05, shape = 16,xlab = 'day relative to onset',
             add = "loess", conf.int = TRUE, 
            add.params = list(color = 'grp', size = 1)) +
  facet_grid(~ fiber_type) +
        scale_colour_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_fill_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_y_sqrt()

ggsave('../data/11_fiber_all_type_daily_intake_onset_grp2.pdf', width = 10, height = 3)
```


```{r}
all_types %>% 
  filter(fdrt %in% -10:30 ) %>% 
  ggscatter(x = 'fdrt', y = 'gram', alpha = 0.05, shape = 16,xlab = 'transplant day',
             add = "loess", conf.int = TRUE, 
            add.params = list(color = 'grp', size = 1)) +
  facet_grid(~ fiber_type) +
    scale_colour_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_fill_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_y_sqrt()

ggsave('../data/11_fiber_all_type_daily_intake_grp3.pdf', width = 10, height = 3)
```


```{r}
# # what is with the inulin a horizontal line
# inulin <- all_types %>% 
#   filter(fiber_type == 'Inulin')
# 
# inulin_same_value <- inulin%>% 
#   count(gram, sort = T) %>% 
#   slice(1) %>% 
#   pull(gram)
# 
# pt_ate <- inulin %>% 
#   filter(gram == inulin_same_value) %>% 
#   count(mrn, sort = T)   
# 
# what_ate <- pt_ate %>% 
#   slice(1) %>% pull(mrn)
# # one banana a day ....
# 
# 
# # what about the Beta-Glucans
# all_types %>% 
#   filter(fiber_type == 'Beta-Glucans') %>% 
#    count(gram, sort = T) %>% 
#   slice(1) %>% 
#   pull(gram)
#  
# # it's because of breakfast cereal 
```

```{r soluble}
all_soluble <- dailycode %>% 
  group_by(mrn, fdrt, soluble) %>% 
  summarise(gram = sum(dcodefiber)) %>% 
  ungroup() %>% 
  filter(!is.na(soluble)) %>% 
  spread('soluble','gram', fill = 0) %>% 
  gather('soluble','gram', names(.)[3]:names(.)[ncol(.)]) %>% 
  inner_join(allpt_gvhd_all %>% distinct(mrn, source, grp))


all_soluble %>% 
  filter(fdrt %in% -10:30) %>% 
  ggscatter(x = 'fdrt', y = 'gram', alpha = 0.05, shape = 16,xlab = 'Transplant day',
             add = "loess", conf.int = TRUE, title = '',palette = 'jco',
            add.params = list(color = 'soluble', size = 1)) +
  scale_y_sqrt() + theme(aspect.ratio = 1)

ggsave('../data/11_fiber_all_soluble_daily_intake.pdf', width = 4, height = 4)


```

```{r}
# Same analysis as above, but including only the No GVHD and LGI patients (joined into one group)

all_soluble %>% 
  filter(fdrt %in% -10:30 & grp != 'Non_LGI') %>% 
  ggscatter(x = 'fdrt', y = 'gram', alpha = 0.05, shape = 16,xlab = 'Transplant day',
             add = "loess", conf.int = TRUE, title = '',palette = 'jco',
            add.params = list(color = 'soluble', size = 1)) +
  scale_y_sqrt() + theme(aspect.ratio = 1)

ggsave('../data/11_fiber_all_soluble_daily_intake_2.pdf', width = 4, height = 4)
```

```{r}
# soluble fiber No GVHD vs. GVHD
all_soluble %>% 
  mutate(gvhd = if_else(grp == 'no_gvhd', 'No', 'Yes'),
         gvhd = factor(gvhd, levels = c('Yes','No'))) %>% 
  filter(fdrt %in% -10:30) %>% 
  ggscatter(x = 'fdrt', y = 'gram', alpha = 0.05, shape = 16,xlab = 'Transplant day',
             add = "loess", conf.int = TRUE, title = 'soluble',
            add.params = list(color = 'gvhd', size = 1)) +
  facet_grid(~ soluble, labeller = 'label_both') +
  scale_y_sqrt()

ggsave('../data/11_fiber_all_soluble_daily_intake_gvhd.pdf', width = 6, height = 4)
```


```{r}
# gvhd groups
all_soluble %>% 
  filter(fdrt %in% -10:30) %>% 
  ggscatter(x = 'fdrt', y = 'gram', alpha = 0.05, shape = 16,xlab = 'Transplant day',
             add = "loess", conf.int = TRUE, title = 'soluble',
            add.params = list(color = 'grp', size = 1)) +
  facet_grid(~ soluble, labeller = 'label_both') +
   scale_colour_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_fill_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_y_sqrt()

ggsave('../data/11_fiber_all_soluble_daily_intake_grp3.pdf', width = 6, height = 4)
```


```{r}
# with only two groups
all_soluble %>% 
  filter(fdrt %in% -10:30 & grp != 'Non_LGI') %>% 
  ggscatter(x = 'fdrt', y = 'gram', alpha = 0.05, shape = 16,xlab = 'Transplant day',
             add = "loess", conf.int = TRUE, title = 'soluble',
            add.params = list(color = 'grp', size = 1)) +
  facet_grid(~ soluble, labeller = 'label_both') +
     scale_colour_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_fill_manual(values = cols, breaks = c("LGI", "no_gvhd", "Non_LGI")) +
  scale_y_sqrt()

ggsave('../data/11_fiber_all_soluble_daily_intake_grp2.pdf', width = 6, height = 4)
```




```{r}
# # maybe break the fdrt down to weeks
# soluble_week <- all_soluble %>% 
#    mutate(timebin = cut_width(fdrt, 7, boundary=0, closed = 'left'))
# 
# soluble_week %>% 
#   filter(d100_a_gvhd %in% c('Y','N')) %>% 
#   ggboxplot(x = 'timebin', y = 'gram', add = 'jitter', add.params = list(alpha = 0.06),
#             xlab = 'transplant day', title = 'Soluble',shape = 16,
#             color = 'd100_a_gvhd') +
#    facet_grid(~ soluble)  +
#   scale_color_manual(values = color_key) +
#   scale_fill_manual(values = color_key) +
#   theme(axis.text.x = element_text(angle = 45, hjust = 1))+
#   scale_y_sqrt()
# 
# # subsample to select the median value for repeated sampling
# soluble_week %>% 
#   filter(d100_a_gvhd %in% c('Y','N')) %>% 
#   group_by(mrn, d100_a_gvhd, soluble, timebin) %>% 
#   summarise(med_week = median(gram)) %>% 
#   ggboxplot(x = 'timebin', y = 'med_week', add = 'jitter', add.params = list(alpha = 0.3),
#             xlab = 'transplant day', title = 'Soluble',shape = 16,
#             color = 'd100_a_gvhd') +
#    facet_grid(~ soluble)  +
#   scale_color_manual(values = color_key) +
#   scale_fill_manual(values = color_key) +
#   theme(axis.text.x = element_text(angle = 45, hjust = 1))+
#   scale_y_sqrt()
# 
# ggsave('../data/patients_16s_data/11_fiber_all_soluble_daily_intake_gvhd_week_median.pdf', width = 8, height = 4)
# ggsave('../data/11_fiber_all_soluble_daily_intake_gvhd_week_median.pdf', width = 8, height = 4)
```
```{r}
# cellulose
# cel <- all_types %>% 
#   filter(fiber_type == 'Cellulose')
# cel %>% 
#   filter(fdrt %in% -30:30 & d100_a_gvhd %in% c('Y','N')) %>% 
#   ggscatter(x = 'fdrt', y = 'gram', alpha = 0.05, shape = 16,xlab = 'transplant day',
#              add = "loess", conf.int = TRUE, title = 'Cellulose',
#             add.params = list(color = 'd100_a_gvhd', size = 1)) +
#   scale_color_manual(values = color_key) +
#   scale_fill_manual(values = color_key)
# 
# ggsave('../data/patients_16s_data/11_fiber_Cellulose_daily_intake_gvhd.pdf', width = 4, height = 4)
# ggsave('../data/11_fiber_Cellulose_daily_intake_gvhd.pdf', width = 4, height = 4)
```


```{r}
# # randomly select one sample per patient 
# set.seed(666)
# 
# type_unisample <- all_types %>% 
#   split(.$fiber_type) %>% 
#   map_dfr(function(df){
#     df %>%  group_by(mrn) %>% sample_n(size = 1, replace = F)
#   })
# 
# 
# type_unisample %>% 
#   filter( d100_a_gvhd %in% c('Y','N')) %>% 
#   ggscatter(x = 'fdrt', y = 'gram', alpha = 0.2, shape = 16,xlab = 'transplant day',
#              add = "loess", conf.int = TRUE, title = 'All fiber types one sample per pt',
#             add.params = list(color = 'd100_a_gvhd', size = 1)) +
#   scale_color_manual(values = color_key) +
#   scale_fill_manual(values = color_key) +
#   scale_y_sqrt() +
#   facet_grid(~ fiber_type)
# 
# ggsave('../data/patients_16s_data/11_fiber_all_daily_intake_gvhd_one_sample_per_pt.pdf', width = 15, height = 4)
# ggsave('../data/11_fiber_all_daily_intake_gvhd_one_sample_per_pt.pdf', width = 15, height = 4)

```

```{r}

# solu_unisample <- all_soluble %>% 
#   split(.$soluble) %>% 
#   map_dfr(function(df){
#     df %>%  group_by(mrn) %>% sample_n(size = 1, replace = F)
#   })
# 
# solu_unisample %>% 
#   filter( d100_a_gvhd %in% c('Y','N')) %>% 
#   ggscatter(x = 'fdrt', y = 'gram', alpha = 0.2, shape = 16,xlab = 'transplant day',
#              add = "loess", conf.int = TRUE, title = 'soluble',
#             add.params = list(color = 'd100_a_gvhd', size = 1)) +
#   facet_grid(~ soluble)  +
#   scale_color_manual(values = color_key) +
#   scale_fill_manual(values = color_key)+
#   scale_y_sqrt()
# 
# ggsave('../data/patients_16s_data/11_fiber_all_soluble_daily_intake_one_sample_per_pt.pdf', width = 6, height = 4)
# ggsave('../data/11_fiber_all_soluble_daily_intake_one_sample_per_pt.pdf', width = 6, height = 4)
```
```{r}
# use the percentage to combat the zero fiber eating thing 
```

```{r}
#median_now <- median(meta$mean_fiber)
meta <- read_csv('~/pipeline/scripts/food_tree/data/090_all_samples_meta_p2d_fg9_dietall_genera90.csv') %>% 
  filter(sdrt %in% -10:30)%>% 
  group_by(mrn) %>% 
  summarise(ave_div = mean(simpson_reciprocal),
            mean_fiber = mean(ave_fiber)) %>% 
  mutate(mean_fiber = mean_fiber * 100) %>% 
  inner_join( allpt_gvhd_all %>% distinct(mrn, source, grp ))

meta  %>% 
  ggscatter(x = 'mean_fiber', y = 'ave_div',
            xlab = 'Prior 2 day fiber intake averaged by patient', ylab = 'microbiome alpha diversity averaged by patient',
            add = "reg.line",  # Add regressin line
           add.params = list(color = "blue", fill = "lightgray"), # Customize line
           conf.int = TRUE, # Add confidence interval
           cor.coef = TRUE, # Add correlation coefficient.
           cor.coeff.args = list(method = "spearman",  label.sep = "\n")) +
  theme(aspect.ratio=1)
ggsave('../data/11_alpha_and_fiber.pdf', width = 10, height = 5)


```

```{r}
median_fiber_intake <- mean(meta$mean_fiber)

grps <- meta %>% 
  mutate(grp = if_else(mean_fiber > median_fiber_intake, 'high', 'low'))

grps %>% 
  ggboxplot(x = 'grp', y = 'ave_div', xlab = 'Compared to average prior 2 day fiber intake',
            ylab = 'Microbiome alpha diversity') +
  stat_compare_means(comparisons= list(c('high', 'low')),
            label= "p.format",
            method= 'wilcox.test',
						correct=FALSE)
ggsave('../data/11-fiber_mean_box.pdf', width = 4, height = 4)
```


```{r}
get_table_from_database('samples_castori_ag')
```


```{r}
ptbg <- allpt_gvhd_all %>% 
  inner_join(allpt_gvhd %>% select(mrn, d100_a_gvhd_onset, hct)) %>% 
  inner_join(samples_castori_ag %>% select(sampleid, datecollection )) %>% 
  mutate(d100_a_gvhd_onset = ymd(d100_a_gvhd_onset),
    sdronset = as.numeric(datecollection - d100_a_gvhd_onset))
ptbg


dailycode %>% 
   inner_join(allpt_gvhd %>% select(mrn, d100_a_gvhd_onset)) %>% 
  inner_join(samples_castori_ag %>% select(sampleid, datecollection )) %>% 
  mutate(d100_a_gvhd_onset = ymd(d100_a_gvhd_onset),
    sdronset = as.numeric(datecollection - d100_a_gvhd_onset))
  filter(fdrt %in% -10:30) %>% 
  group_by(mrn, fdrt, source, grp) %>% 
  summarise(dayfiber = sum(dcodefiber)) %>% 
  ungroup() %>% 
  ggscatter(x = 'fdrt', y = 'dayfiber', add = "loess", conf.int = TRUE, alpha = 0.1, shape = 16,
            add.params = list(color = 'grp'),
            xlab = 'Transplant day', ylab = 'Daily fiber intake' ) +
  scale_y_sqrt() +theme(aspect.ratio = 1)
```

```{r}
meta %>% 
  ggboxplot(x = 'grp', y = 'ave_div', add = 'jitter') +
  stat_compare_means(comparisons= list(c('high', 'low')),
            label= "p.format",
            method= 'wilcox.test',
						correct=FALSE)
ggsave('../data/11_box_low_and_high.pdf')
```
